+++ docs/changelog.rst
@@ -33,6 +33,9 @@ Modifications to existing flavors:
 - Add Kalimantan Utara in  `PROVINCE_CHOICES` for Indonesia local flavor.
   (`gh-385 <https://github.com/django/django-localflavor/pull/385>`_).
 
+- Add validation for women National identity number for Indonesia localflavor.
+  (`gh-386 <https://github.com/django/django-localflavor/pull/386>`_).
+
 Other changes:
 
 - Removed deprecated `generic.checksums.luhn` and `generic.checksums.ean` functions